feat: add html theme for org html exports
Theme is hooked to pre export hook command to inject the css right
before the html is exported. It is inserted inline as css in an org
doc. This also updates org-babel-languages setup to include html and
not just browser. Browser is meant to compile the html and place the
result in the org export which was not intended but very cool.