fix: multiple items including straight, revert config, magit etc.
This commit fixes many issues with the setup for multiple modes
including multiple-cursors where the global list of commands to run
actions for was not being properly set. Magit status was not properly
being initialized so binded the magit status to a global bind. Ensure
auto-revert-mode is enabled globally. Setting to t was not working so
set to 1 instead.